Test::More skip() allows test name
Allow to specify which test name got skipped. This can be the
exact name of a test when skipping only 1 test or a hint about
which group of multiple tests is skippped.

That allows to provide a hint in the printed TAP about the skipped
test, especially when the skip reason is too generic or re-used
multiple times.

The name is just passed through to the test builder which takes care
of handling it, in particular potential undefined values, which
is important for backwards compatibility.